@@ -11,33 +11,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.2.14",
-    "@angular/common": "~11.2.14",
-    "@angular/compiler": "~11.2.14",
-    "@angular/core": "~11.2.14",
-    "@angular/forms": "~11.2.14",
-    "@angular/platform-browser": "~11.2.14",
-    "@angular/platform-browser-dynamic": "~11.2.14",
-    "@angular/router": "~11.2.14",
+    "@angular/animations": "~12.2.11",
+    "@angular/common": "~12.2.11",
+    "@angular/compiler": "~12.2.11",
+    "@angular/core": "~12.2.11",
+    "@angular/forms": "~12.2.11",
+    "@angular/platform-browser": "~12.2.11",
+    "@angular/platform-browser-dynamic": "~12.2.11",
+    "@angular/router": "~12.2.11",
     "bootstrap": "^4.6.0",
     "font-awesome": "^4.7.0",
     "ngx-bootstrap": "^7.1.0",
+    "ngx-toastr": "^14.1.4",
     "rxjs": "~6.6.7",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.13",
-    "@angular/cli": "^11.2.14",
-    "@angular/compiler-cli": "~11.2.14",
-    "@angular/language-service": "~11.2.14",
+    "@angular-devkit/build-angular": "~12.2.11",
+    "@angular/cli": "^12.2.11",
+    "@angular/compiler-cli": "~12.2.11",
+    "@angular/language-service": "~12.2.11",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.3.6",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
@@ -45,6 +46,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.1.6"
+    "typescript": "~4.3.5"
   }
 }
