@@ -1,11 +1,27 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import { NgModule } from "@angular/core";
+import { Routes, RouterModule } from "@angular/router";
+import { NotFoundComponent } from "./core/not-found/not-found.component";
+import { ServerErrorComponent } from "./core/server-error/server-error.component";
+import { TestErrorComponent } from "./core/test-error/test-error.component";
+import { HomeComponent } from "./home/home.component";
+import { ProductDetailsComponent } from "./shop/product-details/product-details.component";
+import { ShopComponent } from "./shop/shop.component";
 
-
-const routes: Routes = [];
+const routes: Routes = [
+  { path: "", component: HomeComponent },
+  { path: "test-error", component: TestErrorComponent },
+  { path: "server-error", component: ServerErrorComponent },
+  { path: "not-found", component: NotFoundComponent },
+  {
+    path: "shop",
+    loadChildren: () =>
+      import("./shop/shop.module").then((mod) => mod.ShopModule),
+  },
+  { path: "**", redirectTo: "", pathMatch: "full" },
+];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes, { relativeLinkResolution: 'legacy' })],
-  exports: [RouterModule]
+  imports: [RouterModule.forRoot(routes)],
+  exports: [RouterModule],
 })
-export class AppRoutingModule { }
+export class AppRoutingModule {}
